+++ R/num_equal.R
@@ -6,6 +6,9 @@ num_equal <- function(x, y, tolerance = .Machine$double.eps ^ 0.5) {
   if (any(is.na(x) != is.na(y))) {
     return(FALSE)
   }
+  if (any(is.nan(x) != is.nan(y))) {
+    return(FALSE)
+  }
 
   attributes(x) <- NULL
   attributes(y) <- NULL