+++ tests/testthat/test-num_equal.R
@@ -29,3 +29,9 @@ test_that("infinite values are handled properly", {
   expect_equal(num_equal(-Inf, Inf), FALSE)
   expect_equal(num_equal(-Inf, Inf, tolerance = 1.e-8), FALSE)
 })
+
+test_that("NaN is different from NA_real_", {
+  expect_false(num_equal(NaN, NA_real_), FALSE)
+  expect_true(num_equal(NaN, NaN))
+  expect_true(num_equal(NA_real_, NA_real_))
+})